remove sidebar slice useTheme hook

+++ src/entities/document/model/hooks/useDeleteDocModal.ts
@@ -1,4 +1,4 @@
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 
 import { useAppDispatch } from "@/shared/hooks/useAppDispatch";
 import { useAppSelector } from "@/shared/hooks/useAppSelector";
@@ -7,21 +7,20 @@ import { deleteDocument, getAllDocuments } from "../api/api";
 import { resetDocument } from "../slice/documentSlice";
 
 export const useDeleteDocModal = () => {
+  const navigate = useNavigate();
   const dispatch = useAppDispatch();
-  const { id } = useParams();
-  const mode = useAppSelector((state) => state.document.mode);
   const { refetch } = getAllDocuments();
   const [deleteDoc] = deleteDocument();
+  const { id } = useParams();
+  const mode = useAppSelector((state) => state.document.mode);
 
   const handleDeleteDoc = async () => {
-    if (mode === "intro") {
-      dispatch(resetDocument());
-    }
-    if (mode === "main") {
-      if (!id) return;
+    if (mode === "main" && id) {
       await deleteDoc(Number(id)).unwrap();
       refetch();
+      navigate("/");
     }
+    dispatch(resetDocument());
   };
   return handleDeleteDoc;
 };
